Account for renaming.
