+++ power/__init__.py
@@ -1,15 +1,16 @@
 from .compute_power import compute_power
 from .metrics import (
-    classification_report,
+    estimate_power_from_accuracy,
+    estimate_power_from_predictions,
     find_dataset_size,
     find_minimum_detectable_effect,
+    make_probability_table,
 )
-from .utils import make_probability_table
 
 __all__ = [
     "compute_power",
-    "classification_report",
-    "find_dataset_size",
+    "estimate_power_from_predictions",
+    "estimate_power_from_accuracy" "find_dataset_size",
     "find_minimum_detectable_effect",
     "make_probability_table",
 ]
